--- a/Dockerfile.production
+++ b/Dockerfile.production
@@ -1,5 +1,5 @@
 # Production Dockerfile for SmartCloudOps AI
-# Multi-stage build for optimized production image
+# Multi-stage build for optimized production image with security hardening
 
 # Build stage
 FROM python:3.11-slim as builder
@@ -10,8 +10,8 @@ ENV PYTHONDONTWRITEBYTECODE=1 \
     PIP_NO_CACHE_DIR=1 \
     PIP_DISABLE_PIP_VERSION_CHECK=1
 
-# Install system dependencies with security updates
-RUN apt-get update && apt-get upgrade -y && apt-get install -y \
+# Install system dependencies (security: no apt-get upgrade to prevent package conflicts)
+RUN apt-get update && apt-get install -y --no-install-recommends \
     gcc \
     g++ \
     libpq-dev \
@@ -23,24 +23,25 @@ RUN python -m venv /opt/venv
 ENV PATH="/opt/venv/bin:$PATH"
 
 # Copy requirements and install Python dependencies
-COPY requirements-production.txt .
-RUN pip install --no-cache-dir -r requirements-production.txt
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
 
 # Production stage
 FROM python:3.11-slim
 
 # Create non-root user for security
-RUN groupadd -r appuser && useradd -r -g appuser appuser
+RUN groupadd -r appuser && useradd -r -g appuser -s /bin/bash -m appuser
 
 # Set environment variables
 ENV PYTHONDONTWRITEBYTECODE=1 \
     PYTHONUNBUFFERED=1 \
     PATH="/opt/venv/bin:$PATH" \
     FLASK_ENV=production \
-    FLASK_APP=app.main:app
+    FLASK_APP=app.main:app \
+    PYTHONPATH="/app"
 
-# Install runtime dependencies with security updates
-RUN apt-get update && apt-get upgrade -y && apt-get install -y \
+# Install runtime dependencies (security: no apt-get upgrade)
+RUN apt-get update && apt-get install -y --no-install-recommends \
     libpq5 \
     curl \
     ca-certificates \
@@ -51,27 +52,33 @@ RUN apt-get update && apt-get upgrade -y && apt-get install -y \
 # Copy virtual environment from builder stage
 COPY --from=builder /opt/venv /opt/venv
 
-# Create application directory
+# Create application directory with proper ownership
 WORKDIR /app
 
 # Copy application code
-COPY app/ ./app/
-COPY ml_models/ ./ml_models/
-COPY configs/ ./configs/
-COPY gunicorn.conf.py .
-COPY start.sh .
+COPY --chown=appuser:appuser app/ ./app/
+COPY --chown=appuser:appuser ml_models/ ./ml_models/
+COPY --chown=appuser:appuser configs/ ./configs/
+COPY --chown=appuser:appuser gunicorn.conf.py .
+COPY --chown=appuser:appuser start.sh .
 
 # Create necessary directories with proper permissions
 RUN mkdir -p /app/logs /app/data /tmp \
     && chown -R appuser:appuser /app /tmp \
-    && chmod +x start.sh
+    && chmod +x start.sh \
+    && chmod 755 /app \
+    && chmod 644 /app/*.py \
+    && chmod 644 /app/*.conf.py
 
 # Security: Remove unnecessary files and set proper permissions
 RUN find /app -type f -name "*.pyc" -delete \
     && find /app -type d -name "__pycache__" -delete \
-    && chmod 755 /app \
-    && chmod 644 /app/*.py \
-    && chmod 644 /app/*.conf.py
+    && find /app -type f -name "*.py" -exec chmod 644 {} \; \
+    && find /app -type d -exec chmod 755 {} \;
+
+# Security: Set read-only root filesystem where possible
+# Note: Application writes only to /tmp, /app/logs, /app/data
+VOLUME ["/app/logs", "/app/data"]
 
 # Switch to non-root user
 USER appuser
@@ -83,9 +90,8 @@ EXPOSE 5000
 HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
     CMD curl -f http://localhost:5000/health || exit 1
 
-# Security: Set read-only root filesystem where possible
-# Note: This requires the application to write only to /tmp, /app/logs, /app/data
-VOLUME ["/app/logs", "/app/data"]
+# Security: Set proper file permissions for runtime
+RUN chmod 755 /app/start.sh
 
 # Start application
 CMD ["gunicorn", "--config", "gunicorn.conf.py", "app.main:app"]