--- a/Dockerfile.production
+++ b/Dockerfile.production
@@ -1,84 +1,93 @@
-# Production Dockerfile for SmartCloudOps AI
-# Multi-stage build for optimized production image with security hardening
+# SmartCloudOps AI - Production Dockerfile
+# Multi-stage build for optimal production image
 
-# Build stage
+# ================================
+# Stage 1: Build Dependencies
+# ================================
 FROM python:3.11-slim as builder
 
-# Set environment variables
-ENV PYTHONDONTWRITEBYTECODE=1 \
-    PYTHONUNBUFFERED=1 \
-    PIP_NO_CACHE_DIR=1 \
-    PIP_DISABLE_PIP_VERSION_CHECK=1
+# Set build arguments
+ARG BUILD_DATE
+ARG VCS_REF
+ARG VERSION
 
-# Install system dependencies (security: no apt-get upgrade to prevent package conflicts)
+# Add metadata labels
+LABEL org.opencontainers.image.created=$BUILD_DATE \
+      org.opencontainers.image.version=$VERSION \
+      org.opencontainers.image.revision=$VCS_REF \
+      org.opencontainers.image.source="https://github.com/smartcloudops/smartcloudops-ai" \
+      org.opencontainers.image.title="SmartCloudOps AI" \
+      org.opencontainers.image.description="AI-powered DevOps platform with anomaly detection and automated remediation" \
+      org.opencontainers.image.vendor="SmartCloudOps" \
+      org.opencontainers.image.licenses="MIT"
+
+# Install system dependencies for building
 RUN apt-get update && apt-get install -y --no-install-recommends \
+    build-essential \
     gcc \
     g++ \
     libpq-dev \
-    && rm -rf /var/lib/apt/lists/* \
-    && apt-get clean
+    && rm -rf /var/lib/apt/lists/*
 
-# Create virtual environment
+# Set work directory
+WORKDIR /build
+
+# Copy requirements first for better caching
+COPY requirements.txt .
+
+# Create virtual environment and install dependencies
 RUN python -m venv /opt/venv
 ENV PATH="/opt/venv/bin:$PATH"
 
-# Copy requirements and install Python dependencies
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+# Upgrade pip and install dependencies
+RUN pip install --no-cache-dir --upgrade pip setuptools wheel && \
+    pip install --no-cache-dir -r requirements.txt
 
-# Production stage
-FROM python:3.11-slim
+# ================================
+# Stage 2: Production Runtime
+# ================================
+FROM python:3.11-slim as production
 
-# Create non-root user for security
-RUN groupadd -r appuser && useradd -r -g appuser -s /bin/bash -m appuser
-
-# Set environment variables
-ENV PYTHONDONTWRITEBYTECODE=1 \
-    PYTHONUNBUFFERED=1 \
-    PATH="/opt/venv/bin:$PATH" \
-    FLASK_ENV=production \
-    FLASK_APP=app.main:app \
-    PYTHONPATH="/app"
-
-# Install runtime dependencies (security: no apt-get upgrade)
+# Install runtime dependencies only
 RUN apt-get update && apt-get install -y --no-install-recommends \
     libpq5 \
     curl \
-    ca-certificates \
     && rm -rf /var/lib/apt/lists/* \
-    && apt-get clean \
-    && update-ca-certificates
+    && apt-get clean
 
-# Copy virtual environment from builder stage
+# Create non-root user
+RUN groupadd -r appgroup && useradd -r -g appgroup -d /app -s /bin/bash appuser
+
+# Copy virtual environment from builder
 COPY --from=builder /opt/venv /opt/venv
+ENV PATH="/opt/venv/bin:$PATH"
 
-# Create application directory with proper ownership
+# Set work directory
 WORKDIR /app
 
 # Copy application code
-COPY --chown=appuser:appuser app/ ./app/
-COPY --chown=appuser:appuser ml_models/ ./ml_models/
-COPY --chown=appuser:appuser configs/ ./configs/
-COPY --chown=appuser:appuser gunicorn.conf.py .
-COPY --chown=appuser:appuser start.sh .
+COPY --chown=appuser:appgroup app/ ./app/
+COPY --chown=appuser:appgroup ml_models/ ./ml_models/
+COPY --chown=appuser:appgroup configs/ ./configs/
+COPY --chown=appuser:appgroup docs/ ./docs/
 
 # Create necessary directories with proper permissions
-RUN mkdir -p /app/logs /app/data /tmp \
-    && chown -R appuser:appuser /app /tmp \
-    && chmod +x start.sh \
-    && chmod 755 /app \
-    && chmod 644 /app/*.py \
-    && chmod 644 /app/*.conf.py
+RUN mkdir -p /app/logs /app/data /app/tmp && \
+    chown -R appuser:appgroup /app
 
-# Security: Remove unnecessary files and set proper permissions
-RUN find /app -type f -name "*.pyc" -delete \
-    && find /app -type d -name "__pycache__" -delete \
-    && find /app -type f -name "*.py" -exec chmod 644 {} \; \
-    && find /app -type d -exec chmod 755 {} \;
+# Set environment variables for production
+ENV FLASK_ENV=production \
+    FLASK_HOST=0.0.0.0 \
+    FLASK_PORT=5000 \
+    PYTHONPATH=/app \
+    PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    PIP_NO_CACHE_DIR=1 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1
 
-# Security: Set read-only root filesystem where possible
-# Note: Application writes only to /tmp, /app/logs, /app/data
-VOLUME ["/app/logs", "/app/data"]
+# Health check
+HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
+    CMD curl -f http://localhost:5000/health || exit 1
 
 # Switch to non-root user
 USER appuser
@@ -86,12 +95,5 @@ USER appuser
 # Expose port
 EXPOSE 5000
 
-# Health check
-HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
-    CMD curl -f http://localhost:5000/health || exit 1
-
-# Security: Set proper file permissions for runtime
-RUN chmod 755 /app/start.sh
-
 # Start application
-CMD ["gunicorn", "--config", "gunicorn.conf.py", "app.main:app"]
+CMD ["python", "app/main.py"]