--- a/smartcloudops-ai/app/anomalies/loading.tsx
+++ b/smartcloudops-ai/app/anomalies/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return null
+}