… .github/workflows/cache-optimization.yml
… .github/workflows/enhanced-pipeline.yml
… .github/workflows/frontend-deploy.yml
… .github/workflows/lighthouse-ci.yml
… .github/workflows/main.yml
… .github/workflows/ml-pipeline.yml
… .github/workflows/performance-testing.yml
… .github/workflows/release-automation.yml
… .pre-commit-config.yaml
… CRITICAL_SECURITY_FIXES_SUMMARY.md
… PRODUCTION_READY_REPORT.md
… SECURITY.md
… alembic/versions/06839fd37ff9_create_anomalies_and_remediation_tables.py
… app/analytics/real_time_dashboard.py
… app/api/anomalies.py
… app/auth.py
… app/auth_module.py
… app/auth_routes.py
… app/chatops/ai_handler.py
… app/chatops/utils.py
… app/chatops_module.py
… app/config.py
… app/database.py
… app/logging_backup/__init__.py
… app/logging_backup/centralized_logging.py
… app/logging_backup/config.py
… app/main.py
… app/ml_module.py
… app/ml_service.py
… app/mlops/__init__.py